Add common ancestor for local web drivers
